requirements.txt
connexion==2.14.2
swagger-ui-bundle==0.0.9
pykafka==2.8.0
Flask==2.2.5
PyYAML==6.0.1
SQLAlchemy==2.0.21
PyMySQL==1.1.0
mysql-connector-python==8.1.0